Ellipsoid calibration visualization: Initial commit.
This plots the ellipsoid fitted points onto two unit circles representing
the upper and lower hemispheres. The point location is the projection from
the 3D unit sphere onto the 2D unit circle, and the color represents the
deviation from the surface.

MATLAB's jet colormap is used to represent the colors. The color blue
indicates that a point is inside the unit sphere, green that it is on the
sphere surface, and red that it is outside the sphere.