UAVSettings: Organizes UAVSettings cache files within the GCS storage path
UAVSettings cache files are saved in the following format:
    <GCS storage path>/boardsettingscache/<board serial>/<timestamp>.uav

Each time a new session is started, the GCS finds the most recent
UAVSettings cache file and compares it to the current UAVSettings. A new
cache file is only saved if the settings differ.

Timestamps are used in the naming of UAVSettings cache files. This requires
the users' computer to have an accurate date and time. This seems like a
reasonable assumption.

Every time the GCS restarts, it arbitrarily rearranges the ordering of XML
attributes. This changes the md5 checksum of a UAVSettings file even if
the settings have not actually changed, causing redundant cache files to
be saved. This issue will not addressed in this feature because there is
an open issue to stop the shuffling of XML attributes.