UAVSettings: Uploads UAVSettings cache files to server.
When a new session is started, and after the current UAVSettings are backed
up locally on the users' machine, The GCS attempts to upload all local
caches to a server via HTTP POST. The local copy of any cache that was
successful uploaded is deleted.

Unlike the directory structure of the local backups, where there is a
directory for every CPU serial number, the server stores all caches in the
same directory using the following naming scheme:
   <CPUSerial>_<timestamp>.uav
This was done because it is not easy to create directories via POST
requests.

It would be ideal to start the cache upload when the users' machine
connects to a new network. However, there is currently functionality for
knowing when an new network connection is established in the GCS.

Currently the server URL is hard coded. Obviously, this is not ideal, but I
didn't find a better place for it.