UAVSettings: Support for creating server directories and retrieving caches
The server was modified to support retrieval of caches from the server and
create directories on the server. This commit contains new and modified
methods for supporting this behavior. When a cache file is uploaded, the
relevant server directory is created if it does not exist. Also, once the
cache is uploaded, it is then retrieved from the server to verify that the
upload was successful.

This commit changes the organization of caches on the server to be the same
as the local cache organization:
	<CPUSerial>/<timestamp>.uav
Each CPUSerial is a directory, and in each CPUSerial directory contain the
pertinent cache files.

This commit also improves the way HTTP POST request was handled, removing
the need of a custom signal to signify when a HTTP request finished.