UAVSettings: GCS compares local and remote cache files before upload
Before the GCS uploads a cache to the server, the GCS downloads the
newest remote cache on the server for the pertinent board. If the MD5
checksum of the newest remote cache matches the MD5 checksum of the local
cache, the cache will not be uploaded because it contains redundant data.