uavobjs: telemetry: change rates
Mostly this is changes from "periodic" to "throttled".  A few objects
that may stay the same for a long time where context is helpful are left
as periodic.  A few rates were changed.

Changeset was tested at this juncture and seemed to work OK talking to
Quanton over USB and bluetooth.